# openapi.yaml
openapi: 3.0.3
info:
  title: x402 Insurance API
  version: 2.3.0
  description: |
    Zero-Knowledge Proof verified insurance against x402 merchant fraud.

    ## Overview
    This API provides micropayment insurance for x402 protocol users. Agents can purchase
    insurance policies to protect against merchant failures, and submit claims with zkp proofs.

    ## What's New in v2.3.0
    - **x402 V2 Payment Flow**: Full facilitator-based verification and settlement
    - **Server-Side Re-fetch Fraud Detection**: `merchant_url` required — server re-fetches to verify claims
    - **Jolt Atlas 3-Argument Prover**: `coverage_amount_units` passed to prover for payout calculation
    - **Claim Response Enrichment**: `server_verified`, `server_http_status`, `merchant_url` in responses
    - **Async Proof Generation**: Use `?async=true` on /claim to avoid timeout risk
    - **Policy Renewal**: Extend policies before expiration with /renew endpoint
    - **Agent Memory Solution**: Lookup policies by wallet address with /policies endpoint

    ## x402 Payment Protocol
    Certain endpoints require x402 payment. The payment flow:
    1. Agent makes request to protected endpoint
    2. Server responds with 402 Payment Required + payment details
    3. Agent signs payment with wallet and resubmits
    4. Server verifies payment and processes request

    ## Pricing
    - Premium: 1% of requested coverage amount
    - Coverage Range: 0.001 - 0.1 USDC (micropayments)
...

...

  contact:
...
servers:
...
paths:
...
  /insure:
    post:
      summary: Create Insurance Policy
      description: |
        Create a new insurance policy. Requires x402 payment of the premium amount.

        **x402 Payment Required:**
        - Premium: percentage-based (1% of requested coverage)
        - Asset: USDC on Base Mainnet
        - Network: eip155:8453 (CAIP-2)

        **Payment Flow (V2):**
        1. POST without payment → 402 response with V2 `accepts` array
        2. Sign payment and encode as base64 JSON
        3. POST with `PAYMENT-SIGNATURE` header → 201 policy created
        4. Server verifies via facilitator and settles on-chain

      tags: [policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
            examples:
              micropayment:
                summary: Micropayment API protection
                value:
                  merchant_url: "https://api.example.com/data"
                  coverage_amount: 0.01
      responses:
        '201':
          description: Policy created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '402':
          description: Payment Required
          headers:
            PAYMENT-REQUIRED:
              schema:
                type: string
              description: x402 V2 payment requirements
            X-Payment-Required:
              schema:
                type: string
              description: x402 V1 payment details (backward compat)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequired'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /claim:
    post:
      summary: Submit Fraud Claim
      description: |
        Submit a fraud claim against a merchant failure. The service will:
        1. Verify the policy is active
        2. Generate zero-knowledge proof of fraud
        3. Verify the proof
        4. Issue USDC refund to policy holder

        **Fraud Detection:**
        Fraud is detected when HTTP response shows:
        - Status code: 503, 500, 404, or other error codes
        - Empty response body
        - Missing critical headers

        **Server-Side Re-fetch (v2.3.0):**
        The server independently re-fetches `merchant_url` to verify the claimed failure.
        The `server_verified` and `server_http_status` fields in the response indicate
        whether the server confirmed the failure.

        **Jolt Prover (v2.3.0):**
        The prover now accepts 3 arguments: `http_status`, `body_length`, `coverage_amount_units`.
        The `coverage_amount_units` is used to compute the payout amount in the proof's public inputs.

        **Async Mode (NEW in v2.2.0):**
        Add `?async=true` query parameter to enable async processing:
        - Returns immediately (202 Accepted) with claim_id
        - Poll GET /claims/{claim_id} for status updates
        - Eliminates timeout risk during proof generation

      tags: [claims]
      parameters:
        - name: async
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Enable async processing (returns 202 immediately)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRequest'
            examples:
              service_unavailable:
                summary: Merchant returned 503
                value:
                  policy_id: "123e4567-e89b-12d3-a456-426614174000"
                  http_response:
                    status: 503
                    body: ""
                    headers:
                      server: "nginx"
      responses:
        '201':
          description: Claim approved and refund issued (sync mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponse'
        '202':
          description: Claim accepted for processing (async mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncClaimResponse'
        '400':
          description: Invalid claim or no fraud detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Proof generation or refund failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /policies:
    get:
      summary: Get Agent Policies
      description: |
        Retrieve all policies for a given agent wallet address.
        Solves the agent memory problem by allowing agents to lookup their policies.

        **Query Parameters:**
        - `wallet`: Required. Ethereum address of the agent

      tags: [policies]
      parameters:
        - name: wallet
          in: query
          required: true
          schema:
            type: string
            pattern: '^0x[a-fA-F0-9]{40}$'
          description: Agent's Ethereum wallet address
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
      responses:
        '200':
          description: List of policies for the wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoliciesResponse'
        '400':
          description: Missing or invalid wallet address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /renew:
    post:
      summary: Renew/Extend Policy
      description: |
        Extend an existing policy before expiration. Requires x402 payment.

        **Features:**
        - Extend by 1-168 hours (max 7 days)
        - Pro-rated fees: (extend_hours / 24) × original_premium
        - Same policy_id preserved
        - Cannot renew expired policies

        **x402 Payment Required:**
        - Amount: Pro-rated renewal fee
        - Asset: USDC on Base Mainnet
        - Network: eip155:8453 (CAIP-2)

        **Payment Flow:**
        1. POST without payment → 402 response with renewal fee
        2. Sign payment with wallet
        3. POST with x402 headers → policy extended

      tags: [policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewPolicyRequest'
            examples:
              renew_24h:
                summary: Extend policy by 24 hours
                value:
                  policy_id: "123e4567-e89b-12d3-a456-426614174000"
                  extend_hours: 24
      responses:
        '200':
          description: Policy renewed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenewPolicyResponse'
        '402':
          description: Payment Required
          headers:
            PAYMENT-REQUIRED:
              schema:
                type: string
              description: x402 V2 payment requirements for renewal fee
            X-Payment-Required:
              schema:
                type: string
              description: x402 V1 payment details (backward compat)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequired'
        '400':
          description: Invalid request or policy expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    CreatePolicyRequest:
      type: object
      required:
        - merchant_url
        - coverage_amount
      properties:
        merchant_url:
          type: string
          format: uri
          description: Merchant API endpoint to protect
          example: "https://api.example.com/data"
        coverage_amount:
          type: number
          minimum: 0.001
          maximum: 0.1
          description: Coverage amount in USDC
          example: 0.01
    PolicyResponse:
      type: object
      properties:
        policy_id:
          type: string
          format: uuid
        agent_address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
          description: Ethereum address of policy holder
        coverage_amount:
          type: number
        premium:
          type: number
        status:
          type: string
          enum: [active, claimed, expired]
        expires_at:
          type: string
          format: date-time
    Policy:
      allOf:
        - $ref: '#/components/schemas/PolicyResponse'
        - type: object
          properties:
            merchant_url:
              type: string
            merchant_url_hash:
              type: string
            created_at:
              type: string
              format: date-time
    ClaimRequest:
      type: object
      required:
        - policy_id
        - http_response
        - merchant_url
      properties:
        merchant_url:
          type: string
          format: uri
          description: Merchant API URL (must match policy). Used for server-side re-fetch fraud detection.
          example: "https://api.example.com/data"
        policy_id:
          type: string
          format: uuid
        http_response:
          type: object
          required:
            - status
            - body
          properties:
            status:
              type: integer
              description: HTTP status code
              example: 503
            body:
              type: string
              description: Response body (empty for fraud)
              example: ""
            headers:
              type: object
              additionalProperties:
                type: string
              description: HTTP headers
              example:
                server: "nginx"
    ClaimResponse:
      type: object
      properties:
        claim_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        proof:
          type: string
          description: Zero-knowledge proof (hex)
          pattern: '^0x[a-fA-F0-9]+$'
        public_inputs:
          type: array
          items:
            type: integer
          description: |
            Public inputs to the proof:
            [0]: is_fraud (1 = yes, 0 = no)
            [1]: http_status
            [2]: body_length
            [3]: suggested_payout
        payout_amount:
          type: number
          description: Actual payout in USDC
        refund_tx_hash:
          type: string
          pattern: '^0x[a-fA-F0-9]{64}$'
          description: Blockchain transaction hash
        status:
          type: string
          enum: [paid, pending, rejected]
        server_verified:
          type: boolean
          description: Whether server independently re-fetched the merchant URL to verify failure
        server_http_status:
          type: integer
          description: HTTP status code from server's independent re-fetch of merchant URL
        merchant_url:
          type: string
          format: uri
          description: The merchant API URL that was checked
        proof_url:
          type: string
          description: URL to retrieve full proof data
    AsyncClaimResponse:
      type: object
      properties:
        claim_id:
          type: string
          format: uuid
          description: Claim ID for polling status
        status:
          type: string
          enum: [processing]
          description: Initial status (always "processing")
        poll_url:
          type: string
          description: URL to poll for claim status
          example: "/claims/123e4567-e89b-12d3-a456-426614174000"
        estimated_completion_seconds:
          type: integer
          description: Estimated time to completion
          example: 20
...
    PoliciesResponse:
      type: object
      properties:
        wallet:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
          description: Queried wallet address
        total_policies:
          type: integer
          description: Total number of policies for this wallet
        active_policies:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
          description: Currently active policies
        claimed_policies:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
          description: Policies that have been claimed
        expired_policies:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
          description: Expired policies
    RenewPolicyRequest:
      type: object
      required:
        - policy_id
        - extend_hours
      properties:
        policy_id:
          type: string
          format: uuid
          description: ID of the policy to renew
        extend_hours:
          type: integer
          minimum: 1
          maximum: 168
          description: Number of hours to extend (1-168 = 7 days)
          example: 24
    RenewPolicyResponse:
      type: object
      properties:
        policy_id:
          type: string
          format: uuid
        old_expires_at:
          type: string
          format: date-time
          description: Previous expiration time
        new_expires_at:
          type: string
          format: date-time
          description: New expiration time after renewal
        extend_hours:
          type: integer
          description: Number of hours extended
        renewal_fee:
          type: number
          description: Pro-rated renewal fee paid
        renewal_count:
          type: integer
          description: Total number of times this policy has been renewed
        total_renewal_fees:
          type: number
          description: Total renewal fees paid for this policy
        status:
          type: string
          enum: [active]
...
    PaymentRequired:
      type: object
      properties:
        x402Version:
          type: integer
          example: 2
        accepts:
          type: array
          items:
            type: object
            properties:
              scheme:
                type: string
                example: "exact"
              network:
                type: string
                example: "eip155:8453"
                description: "CAIP-2 network identifier"
              amount:
                type: string
                description: Amount in smallest units (USDC has 6 decimals)
                example: "1000"
              asset:
                type: string
                example: "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913"
                description: "USDC contract address"
              payTo:
                type: string
                description: Backend wallet address
              maxTimeoutSeconds:
                type: integer
                example: 60
              extra:
                type: object
        resource:
          type: object
          properties:
            url:
              type: string
              example: "/insure"
            method:
              type: string
              example: "POST"
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message